Race condition on connection fix:
Problem

The connect callback method can be invoked before the  RouterClient constructor method has completed – yes it does happen it! The initialHandshakeResponse - comes back in (on another thread – intitiated by the handshake request)- while still in constructor - and connectHandler is called - and the handler that passed in has reference to FinSemble.RouterClient that is still null as its constructor has not yet returned and is passed in to the constructors of the other components- then they throw given null reference error.

Solution:
Move logic from  RouterClient Constructor in to  separate Init() method
And call Init from FinSemble after RouterCLient created.
Also put the callback in its own method.